Updated this script to account for changes in the database M    database/mtams.insert.app.sql
Use this as a reference to the order that the insert statments should be run
to statisfy FK constrataints
A    database/table.order.lst

A linux bash script to take a list of insert statements and order them accourding to the the above list
AM   database/insert.order.sh

An insert template file, with insert statements for each table, and very basic example data. These insert will need to be modified to be usefull in this application.
A    database/mtams.insert.template.sql